requirements.txt
pywebview==4.4.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyinstaller==6.3.0
